extractor/requirements.txt
pydantic==2.11.3
boto3==1.38.3
openai==1.76.0
beautifulsoup4==4.13.4
requests==2.32.3
python-dotenv==1.1.0
